@@ -1,6 +1,6 @@
 project(pixels-cli)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 
 include(ExternalProject)
 include(ProcessorCount)
