@@ -36,10 +36,10 @@ public:
   void close() override;
 
   void read(std::shared_ptr <ByteBuffer> input,
-            pixels::proto::ColumnEncoding &encoding,
+            const pixels::fb::ColumnEncoding* encoding,
             int offset, int size, int pixelStride,
             int vectorIndex, std::shared_ptr <ColumnVector> vector,
-            pixels::proto::ColumnChunkIndex &chunkIndex,
+            const pixels::fb::ColumnChunkIndex* chunkIndex,
             std::shared_ptr <PixelsBitMask> filterMask) override;
 
 private:
