@@ -27,6 +27,7 @@
 
 #include "reader/ColumnReader.h"
 #include "encoding/RunLenIntDecoder.h"
+#include "pixels_generated.h"
 
 class StringColumnReader : public ColumnReader
 {
@@ -38,10 +39,10 @@ public:
     void close() override;
 
     void read(std::shared_ptr <ByteBuffer> input,
-              pixels::proto::ColumnEncoding &encoding,
+              const pixels::fb::ColumnEncoding* encoding,
               int offset, int size, int pixelStride,
               int vectorIndex, std::shared_ptr <ColumnVector> vector,
-              pixels::proto::ColumnChunkIndex &chunkIndex,
+              const pixels::fb::ColumnChunkIndex* chunkIndex,
               std::shared_ptr <PixelsBitMask> filterMask) override;
 
 private:
@@ -74,6 +75,6 @@ private:
      * In this method, we have reduced most of significant memory copies.
      */
     void readContent(std::shared_ptr <ByteBuffer> input,
-                     uint32_t inputLength, pixels::proto::ColumnEncoding &encoding);
+                     uint32_t inputLength, const pixels::fb::ColumnEncoding* encoding);
 };
 #endif //PIXELS_STRINGCOLUMNREADER_H
