@@ -40,7 +40,6 @@ LongColumnWriter::LongColumnWriter(
 
 int LongColumnWriter::write(std::shared_ptr<ColumnVector> vector, int size)
 {
-  std::cout << "In LongColumnWriter" << std::endl;
   auto columnVector = std::static_pointer_cast<LongColumnVector>(vector);
   if (!columnVector)
   {
@@ -125,7 +124,7 @@ void LongColumnWriter::newPixel()
   if (runlengthEncoding)
   {
     std::vector<byte> buffer(curPixelVectorIndex *
-    sizeof(int));
+    sizeof(long));
     int resLen;
     encoder->encode(curPixelVector.data(), buffer.data(), curPixelVectorIndex,
                     resLen);
@@ -136,20 +135,20 @@ void LongColumnWriter::newPixel()
     EncodingUtils encodingUtils;
 
     curVecPartitionBuffer =
-        std::make_shared<ByteBuffer>(curPixelVectorIndex * sizeof(int));
+        std::make_shared<ByteBuffer>(curPixelVectorIndex * sizeof(long));
     if (byteOrder == ByteOrder::PIXELS_LITTLE_ENDIAN)
     {
       for (int i = 0; i < curPixelVectorIndex; i++)
       {
         encodingUtils.writeLongLE(curVecPartitionBuffer,
-                                  (int) curPixelVector[i]);
+                                  (int64_t) curPixelVector[i]);
       }
     } else
     {
       for (int i = 0; i < curPixelVectorIndex; i++)
       {
         encodingUtils.writeLongBE(curVecPartitionBuffer,
-                                  (int) curPixelVector[i]);
+                                  (int64_t) curPixelVector[i]);
       }
     }
 
@@ -160,17 +159,14 @@ void LongColumnWriter::newPixel()
   ColumnWriter::newPixel();
 }
 
-pixels::proto::ColumnEncoding LongColumnWriter::getColumnChunkEncoding() const
+const flatbuffers::Offset<pixels::fb::ColumnEncoding> LongColumnWriter::getColumnChunkEncoding(flatbuffers::FlatBufferBuilder& fbb) const
 {
-  pixels::proto::ColumnEncoding columnEncoding;
+
   if (runlengthEncoding)
   {
-    columnEncoding.set_kind(
-        pixels::proto::ColumnEncoding::Kind::ColumnEncoding_Kind_RUNLENGTH);
+    return pixels::fb::CreateColumnEncoding(fbb,pixels::fb::EncodingKind_RUNLENGTH);
   } else
   {
-    columnEncoding.set_kind(
-        pixels::proto::ColumnEncoding::Kind::ColumnEncoding_Kind_NONE);
+    return pixels::fb::CreateColumnEncoding(fbb,pixels::fb::EncodingKind_NONE);
   }
-  return columnEncoding;
 }
